全部迁移，开始差错

+++ AssetReader/Brotli/IntReader.cs
@@ -23,8 +23,8 @@ namespace Org.Brotli.Dec {
         /// </remarks>
         internal static void Convert(IntReader ir, int intLen) {
             for (var i = 0; i < intLen; ++i) {
-                ir.intBuffer[i] = ((ir.byteBuffer[i * 4] & unchecked((int)(0xFF)))) | ((ir.byteBuffer[(i * 4) + 1] & unchecked((int)(0xFF))) << 8) | ((ir.byteBuffer[(i * 4) + 2] & unchecked((int)(0xFF))) << 16) | ((ir.byteBuffer[(i * 4) + 3] & unchecked((int
-                    )(0xFF))) << 24);
+                ir.intBuffer[i] = ((ir.byteBuffer[i * 4] & unchecked(0xFF))) | ((ir.byteBuffer[(i * 4) + 1] & unchecked(0xFF)) << 8) | ((ir.byteBuffer[(i * 4) + 2] & unchecked(0xFF)) << 16) | ((ir.byteBuffer[(i * 4) + 3] & unchecked(
+                    0xFF)) << 24);
             }
         }
     }
